New version of tuple_cat inspired by ideas from Stephan T. Lavavej of Microsoft
Stephan suggested using a meta tuple and exploding it in one shot rather
the recursive the earlier implementation.

This implementation looks similar to the one STL wrote for MS STL, but
it is a slightly different take.

The to_tuple template is almost verbatim copy from the VS 2015 tuple
file.